3.11 Update
Updated GClasses, rewrote some code to make the mod work for 3.11
