+++ Headbob.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.4.33122.133
+VisualStudioVersion = 17.10.34928.147
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Headbob", "Headbob.csproj", "{5C4E21CF-F235-4CFE-B085-94C6F1126FF0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Headbob", "Headbob.csproj", "{9076ADE2-F4EE-4B5D-B9A6-29A054CF0A7E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5C4E21CF-F235-4CFE-B085-94C6F1126FF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5C4E21CF-F235-4CFE-B085-94C6F1126FF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5C4E21CF-F235-4CFE-B085-94C6F1126FF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5C4E21CF-F235-4CFE-B085-94C6F1126FF0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9076ADE2-F4EE-4B5D-B9A6-29A054CF0A7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9076ADE2-F4EE-4B5D-B9A6-29A054CF0A7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9076ADE2-F4EE-4B5D-B9A6-29A054CF0A7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9076ADE2-F4EE-4B5D-B9A6-29A054CF0A7E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A4C83A5C-D5DB-4ACE-BD92-9519A5865171}
+		SolutionGuid = {09F9952E-ECA7-4B79-A94D-0FBCDA1BC1B4}
 	EndGlobalSection
 EndGlobal
