+++ Plugin.cs
@@ -1,15 +1,29 @@
 ﻿using BepInEx;
+using BepInEx.Logging;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Headbob.ExamplePatches;
 
 namespace Headbob
 {
-    [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
+    // first string below is your plugin's GUID, it MUST be unique to any other mod. Read more about it in BepInEx docs. Be sure to update it if you copy this project.
+    [BepInPlugin("Headbob.UniqueGUID", "Headbob", "1.0.0")]
     public class Plugin : BaseUnityPlugin
     {
+        public static ManualLogSource LogSource;
+
+        // BaseUnityPlugin inherits MonoBehaviour, so you can use base unity functions like Awake() and Update()
         private void Awake()
         {
-            new HeadBobRangePatch().Enable();
-            new HeadBobClampPatch().Enable();
-            Logger.LogInfo($"Plugin {PluginInfo.PLUGIN_GUID} is loaded!");
+            // save the Logger to variable so we can use it elsewhere in the project
+            LogSource = Logger;
+            LogSource.LogInfo("plugin loaded!");
+            // uncomment line(s) below to enable desired example patch, then press F6 to build the project:
+            new HeadBobRange().Enable();
+            new HeadBobPatch().Enable();
         }
     }
 }
